Update routes tree
