+++ packages/brain/src/modules/apiServers/launchpad/routes/feeRecipients/index.ts
@@ -1,0 +1,3 @@
+import feeRecipientsRouter from "./route.js";
+
+export { feeRecipientsRouter };
