+++ packages/brain/src/modules/apiServers/launchpad/routes/keystores/index.ts
@@ -1,0 +1,3 @@
+import keystoresRouter from './route.js';
+
+export { keystoresRouter };
