+++ packages/brain/src/modules/apiServers/launchpad/routes/keystores/validation.ts
@@ -1,5 +1,5 @@
-import { tags as availableTags, isValidBlsPubkey, isValidEcdsaPubkey, Tag, Web3signerDeleteRequest } from "@stakingbrain/common";
-import { BrainKeystoreImportRequest, BrainPubkeysFeeRecipients } from "../types.js";
+import { Web3signerDeleteRequest, isValidBlsPubkey } from "@stakingbrain/common";
+import { BrainKeystoreImportRequest } from "../../types.js";
 
 export function validateImportKeystoresRequestBody(request: BrainKeystoreImportRequest): void {
     // Check if the request contains keystores
@@ -43,31 +43,4 @@ export function validateDeleteRequestBody(deleteReq: Web3signerDeleteRequest): v
             throw new Error(`Invalid pubkey format: ${invalidPubkeys.join(", ")}. Pubkeys should follow BLS format (beginning with 0x)`);
         }
     }
-}
-
-export function validatePubkeysQueryParam(pubkeys: string | string[] | undefined): void {
-    if (!pubkeys) return;
-
-    const pubkeyArray = Array.isArray(pubkeys) ? pubkeys : pubkeys.split(',');
-    const invalidPubkeys = pubkeyArray.filter(pubkey => !isValidBlsPubkey(pubkey));
-
-    if (invalidPubkeys.length > 0)
-        throw new Error(`Invalid pubkey format: ${invalidPubkeys.join(", ")}. Pubkeys should follow BLS format (beginning with 0x)`);
-}
-
-export function validateUpdateFeeRecipientRequestBody(requestBody: BrainPubkeysFeeRecipients): void {
-    if (!requestBody || !Array.isArray(requestBody.validators)) {
-        throw new Error("The request body must contain a 'validators' array.");
-    }
-
-    const { validators } = requestBody;
-    validators.forEach((validator, index) => {
-        if (typeof validator.pubkey !== 'string' || !isValidBlsPubkey(validator.pubkey)) {
-            throw new Error(`Validator at index ${index} is missing a valid or has an invalid BLS public key: ${validator.pubkey}.`);
-        }
-
-        if (typeof validator.feeRecipient !== 'string' || !isValidEcdsaPubkey(validator.feeRecipient)) {
-            throw new Error(`Validator at index ${index} is missing a valid or has an invalid Ethereum address: ${validator.feeRecipient}.`);
-        }
-    });
 }
